requirements.txt
# Langgraph 和 Langchain
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0

# 向量存储
chromadb>=0.5.0

# 网页搜索
tavily-python>=0.5.0

# 网页抓取
requests>=2.31.0
beautifulsoup4>=4.12.0
markdownify>=0.11.6

# BM25 检索
rank-bm25>=0.2.2

# 中文分词
jieba>=0.42.1

# Markdown 处理
markdown>=3.5.0

# OpenAI API
openai>=1.0.0

# 阿里云 DashScope API
dashscope>=1.0.0

# 其他
python-dotenv>=1.0.0
